Initial ugly ChannelPreview shows on mobile
