+++ packages/app/features/chat/ChannelList.tsx
@@ -3,19 +3,8 @@ import { useRef } from 'react'
 import { Image, StyleSheet, View } from 'react-native'
 import { Text } from '@my/ui'
 import { FlashList, ListRenderItemInfo } from '@shopify/flash-list'
+import { ChannelPreview } from './ChannelPreview'
 import { useChannels } from './useChannels'
-
-function isValidImageUrl(url: string): boolean {
-  // TODO: this needs work
-  const urlRegex = /^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/?#[\]@!\$&'\(\)\*\+,;=.]+$/
-  if (urlRegex.test(url)) {
-    console.log('VALID:', url)
-    return true
-  } else {
-    console.log('INVALID:', url)
-    return false
-  }
-}
 
 export const ChannelList = () => {
   const channels = useChannels()
@@ -24,16 +13,12 @@ export const ChannelList = () => {
   const renderItem = ({ index }: ListRenderItemInfo<Channel>) => {
     const channel = channels[index]
     if (!channel) return <></>
-    const imageUri = isValidImageUrl(channel.metadata.picture)
-      ? channel.metadata.picture
-      : 'https://placekitten.com/100/100'
-    return (
-      <View style={styles.container}>
-        <Text color="$moonRaker">{channel.metadata.name}</Text>
-        <Text color="$moonRaker">{channel.metadata.about}</Text>
 
-        <Image source={{ uri: imageUri }} style={styles.avatar} />
-      </View>
+    return (
+      <ChannelPreview
+        channel={channel}
+        onPress={() => console.log(`Clicked channel: ${channel.metadata.name}`)}
+      />
     )
   }
